PC hosts discovery implemented

+++ Model/ArpRecord.cs
@@ -7,13 +7,13 @@ using System.Net;
 /// </summary>
 public record ArpRecord
 {
-    public ArpRecord(UInt32 mac, IPAddress ip)
+    public ArpRecord(UInt64 mac, IPAddress ip)
     {
         Mac = mac;
         Ip = ip;
     }
 
-    public UInt32 Mac { get; private set; }
+    public UInt64 Mac { get; private set; }
     public IPAddress? Ip { get; private set; }
 }
